Rename modules and remove dead code
