+++ src/modules/common/IdModule.tsx
@@ -1,6 +1,6 @@
 import {v4 as uuidv4} from 'uuid';
 
-export const Id = {
+export const IdModule = {
     generate
 }
